rename file to match new function name
